+++ spring-project/src/main/java/com/c108/springproject/order/dto/request/OrderUpdateReqDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderUpdateReqDto {
-    private int orderNo;          // 수정할 주문 번호
+    private Long orderNo;          // 수정할 주문 번호
     private String requests;      // 배송 요청사항 변경
     private int orderCategoryNo; // 주문 상태 변경
     private int delCategoryNo;   // 배송 상태 변경
